Contrail to support 4 byte AS
Contrail only supports 2 byte AS today. There is a requirement to extend
it to use 4 byte AS also.
This is first draft to support the same. Some testcaes are added and
some more may need to be added.

Partial-Bug: #1416551

+++ src/bgp/routing-policy/routing_policy_action.h
@@ -81,16 +81,16 @@ public:
 
 class UpdateAsPath : public RoutingPolicyUpdateAction {
 public:
-    UpdateAsPath(const std::vector<uint16_t> &asn_list);
+    UpdateAsPath(const std::vector<uint32_t> &asn_list);
     virtual ~UpdateAsPath() {}
 
     virtual void operator()(BgpAttr *out_attr) const;
     std::string ToString() const;
     virtual bool IsEqual(const RoutingPolicyAction &as_path) const;
-    const std::vector<uint16_t> &asn_list() const { return asn_list_; }
+    const std::vector<uint32_t> &asn_list() const { return asn_list_; }
 
 private:
-    std::vector<uint16_t> asn_list_;
+    std::vector<uint32_t> asn_list_;
 };
 
 class UpdateCommunity : public RoutingPolicyUpdateAction {
